Adding scheduled task for reservation updates with Coravel library
If reservation starts in less than 2 days then concrete car is assigned to reservation.
For reservations that starts in more than 2 days only car model is assigned wich avoid blocking concrete car for long period of time.

Scheduler runs every day at 2AM and checks for reservations starting in next 2 days without assigned concrete car.
If it found such reservation it attempts to assign concrete car with model and location given in reservation.
If it not found any in given location that matches with reserved model it logs and save the issuse in database.
Then admin can get info about issue and decide what to do: physical relocate cars or send proposal to user to change model
or start location or even cancel the reservation.